requirements.txt
flask
mysql-connector-python
pandas
openpyxl
pdfplumber
sentence-transformers
langchain
faiss-cpu
apscheduler

